+++ protected/humhub/docs/CHANGELOG.md
@@ -6,13 +6,13 @@ HumHub Change Log
 
 - Fix: Serialization of notifications without originator fails
 - Fix: Hide unapproved member activities
+- Fix #3313: Unable to deny invitation to private space
 
 
 1.3.4  (September 25, 2018)
 ---------------------------
 
 - Fix: Theme parent lookup cache causes installer crash
-- Fix #3313: Unable to deny invitation to private space
 
 
 1.3.3  (September 24, 2018)